Move user service to /services
